# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_vmarble_warehouse-management-service_internal_domain.Money:
    properties:
      amount:
        type: integer
      currency:
        type: string
    type: object
...
  github_com_vmarble_warehouse-management-service_internal_platform_httpkit.PagedResult-internal_module_order_PO:
    properties:
      current_page:
        type: integer
      items:
        items:
          $ref: '#/definitions/internal_module_order.PO'
        type: array
      limit:
        type: integer
      total_is_estimate:
        type: boolean
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  internal_module_order.CreateLineItemInput:
    properties:
      quantity:
        type: integer
      selling_price:
        $ref: '#/definitions/github_com_vmarble_warehouse-management-service_internal_domain.Money'
      sku_id:
        type: string
    type: object
  internal_module_order.CreatePOInput:
    properties:
      code:
        type: string
      expected_delivery:
        type: string
      line_items:
        items:
          $ref: '#/definitions/internal_module_order.CreateLineItemInput'
        type: array
    type: object
  internal_module_order.LineItem:
    properties:
      id:
        type: string
      po_id:
        type: string
      quantity:
        type: integer
      selling_price:
        $ref: '#/definitions/github_com_vmarble_warehouse-management-service_internal_domain.Money'
      sku_id:
        type: string
    type: object
  internal_module_order.PO:
    properties:
      code:
        type: string
      created_at:
        type: string
      expected_delivery:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      item_count:
        type: integer
      line_items:
        items:
          $ref: '#/definitions/internal_module_order.LineItem'
        type: array
      total_quantity:
        type: integer
      total_skus:
        type: integer
    type: object
...
info:
  contact: {}
  description: Backend API for warehouse & production management.
  title: VMARBLE Warehouse Management Service API
  version: 0.1.0
paths:
...
  /api/v1/pos:
    get:
      parameters:
      - description: page number (default 1)
        in: query
        name: page
        type: integer
      - description: items per page (default 10, max 100)
        in: query
        name: limit
        type: integer
      - description: search by PO code (ILIKE)
        in: query
        name: search
        type: string
      - description: 'sort column: code, expected_delivery (default created_at)'
        in: query
        name: sort_by
        type: string
      - description: 'sort direction: asc, desc (default desc)'
        in: query
        name: order
        type: string
      - description: filter created_at >= from (YYYY-MM-DD or RFC3339, Asia/Ho_Chi_Minh
          local)
        in: query
        name: from
        type: string
      - description: filter created_at < to + 1 day (YYYY-MM-DD or RFC3339, Asia/Ho_Chi_Minh
          local; inclusive)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_vmarble_warehouse-management-service_internal_platform_httpkit.PagedResult-internal_module_order_PO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List POs
      tags:
      - order
    post:
      consumes:
      - application/json
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_module_order.CreatePOInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_module_order.PO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create PO
      tags:
      - order
  /api/v1/pos/{id}:
    delete:
      parameters:
      - description: po id (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Void PO (soft delete)
      tags:
      - order
    get:
      parameters:
      - description: po id (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_module_order.PO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get PO
      tags:
      - order
  /api/v1/pos/{id}/line-items:
    get:
      parameters:
      - description: po id (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_module_order.LineItem'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List PO line items
      tags:
      - order
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"